Moved alert slack message builder into alert
