+++ elementary/clients/slack/schema.py
@@ -9,14 +9,3 @@ class SlackMessageSchema(BaseModel):
     text: Optional[str] = None
     attachments: Optional[list] = None
     blocks: Optional[list] = None
-
-
-class AlertDetailsPartSlackMessageSchema(BaseModel):
-    result: Optional[SlackBlocksType] = None
-    configuration: Optional[SlackBlocksType] = None
-
-
-class AlertSlackMessageSchema(BaseModel):
-    title: Optional[SlackBlocksType] = None
-    preview: Optional[SlackBlocksType] = None
-    details: Optional[AlertDetailsPartSlackMessageSchema] = None
