+++ elementary/monitor/alerts/schema/slack_alert.py
@@ -1,0 +1,15 @@
+from typing import Optional
+from pydantic import BaseModel
+
+from elementary.clients.slack.schema import SlackBlocksType
+
+
+class AlertDetailsPartSlackMessageSchema(BaseModel):
+    result: Optional[SlackBlocksType] = None
+    configuration: Optional[SlackBlocksType] = None
+
+
+class AlertSlackMessageSchema(BaseModel):
+    title: Optional[SlackBlocksType] = None
+    preview: Optional[SlackBlocksType] = None
+    details: Optional[AlertDetailsPartSlackMessageSchema] = None
